fix: complete Wan MXFP8 review follow-ups
